Add matrix avx multiplication, need to fix errors in printf

+++ 1st-sem/matrix-parallel/matrix.h
@@ -2,7 +2,7 @@
 
 struct matrix_t {
     
-    double** data; 
+    float** data; 
     int rows;
     int cols;
 
@@ -14,7 +14,7 @@ int matrix_init(matrix_t* matrix, int rows, int cols);
 
 int matrix_dump(matrix_t* matrix);
 
-int matrix_fill(matrix_t* matrix, double* data, int size);
+int matrix_fill(matrix_t* matrix, float* data, int size);
 
 matrix_t matrix_add(matrix_t* lhs, matrix_t* rhs);
